🔧(front) make the package usable as a dep
There were various issues like wrong extension of the js files. Also
I export the cunningham.ts content in order for the consumers to
manipulate it directly instead of it being magically bundled inside
the style. That's why library.css has been created, it does not
include the cunningham tokens, it is up to the consumers, in case they
want to make some customization.