✨(i18n) allow custom locales
We faced an issue: it is not possible to import cunningham locales
directly because of the lack of an export. Knowing we want to add
our custom translations to it, I added this approach to import the
local to the src folder and then a custom CunninghamProvider than
just inject the custom translation into the real CunninghamProvider.
The custom translations of this package should be put inside
src/locales/en-US.json or fr.json.