⬆️(deps) make react a peer dependencies
Currently `react` and `react-dom` are declared as deps for this package.
This is weird as it can create deps conflict when a project installed its own
version of those package and this one is different that the one installed in
this package.

So as a workaround, we declare `react` and `react-dom` as peer deps with
a minimal requiremen of version 19 (as cunningham requires at least
this version) then we install those package as devDependencies to be able
to work in this repo.