♻️(opti) remove base64 fonts and roboto from styles
Previously the fonts were inlined as base64 in the style file, which
made the file heavy, which is not what we want. Instead we delegate
the font loading to the consumer app, and we provide a built-in fonts.scss
to import on consummer apps in order to pull the fonts as static also.
Also remove Roboto font imported from Cunningham.