🔧(tree) add dndRootElement prop to TreeView
If dndRootElement in not defined, the drag and drop
events listen the full document. When a drag
start, an invisible overlay is created to capture
the drop events. This overlay can block the dnd
of other elements in the page, like the
DND of Blocknote editor in our case.
By setting the dndRootElement, the overlay will
be confined to the specified element,
preventing it from blocking DND events
of other elements.