♻️(layout) remove zindex on layout header
Now that we removed in the previous commit the zindex of the modal,
the header was diplayed on top of the modal. Removing it allow to
revert back to the default stacking context of the DOM. We will let
consumer apps do whatever they want in they need to overwrite it.