✨(tree) add custom row props to TreeView components
We need to be able to pass custom props to the
rows of the TreeView component for better
flexibility, particulary about accessibility. This
commit introduces a new prop `rowProps` to the
TreeView component that allows users to specify
additional HTML attributes for each row.
We added as well itemProps to TreeViewItem for
the same purpose.